refactor: get week weather from global state.
