+++ src/components/week-weather/week-weather.tsx
@@ -1,13 +1,18 @@
 import WeatherCard from './weather-card/weather-card';
 
+import { RootState } from '../../store';
+import { useSelector, useDispatch } from 'react-redux';
+
 import styles from './week-weather.module.scss';
-import ConsolidatedWeather from '../../models/consolidated_weather';
+import { ConsolidatedWeather } from '../../models';
 
 interface Props {
   weekWeather: ConsolidatedWeather[];
 }
 
-function WeekWeather({ weekWeather }: Props) {
+function WeekWeather() {
+  const { weekWeather } = useSelector((state: RootState) => state.weather);
+
   return (
     <section className={styles.container}>
       {weekWeather &&
